+++ backend/src/main/java/org/launchcode/Amethyst/entity/CartItem.java
@@ -4,7 +4,7 @@ import jakarta.persistence.*;
 
 import java.sql.Time;
 
-@Entity
+//@Entity
 public class CartItem {
 
     @GeneratedValue
