+++ backend/src/main/java/org/launchcode/Amethyst/services/CartService.java
@@ -1,0 +1,9 @@
+package org.launchcode.Amethyst.services;
+
+import org.launchcode.Amethyst.dto.CartDto;
+
+public interface CartService {
+
+    CartDto createCart(CartDto cartDto);
+    CartDto getCartById(int id);
+}
